+++ cores/arduino/HardwareSerial.h
@@ -47,7 +47,7 @@ class HardwareSerial : public Stream
     virtual int available(void) = 0;
     virtual int peek(void) = 0;
     virtual int read(void) = 0;
-    virtual size_t read(uint8_t *buffer, size_t size) = 0;
+    virtual int read(uint8_t *buffer, size_t size) = 0;
     virtual void flush(void) = 0;
     virtual size_t write(uint8_t) = 0;
     using Print::write; // pull in write(str) and write(buf, size) from Print
