+++ libraries/Wire/src/Wire.h
@@ -52,7 +52,7 @@ public:
 
     virtual int available(void);
     virtual int read(void);
-    virtual size_t read(uint8_t *buffer, size_t size);
+    virtual int read(uint8_t *buffer, size_t size);
     virtual int peek(void);
     virtual void flush(void);
     void onReceive(void(*)(int));
